Перенос в отдельный файл
